+++ src/index.ts
@@ -21,6 +21,7 @@
 
 export { InruptClientError } from "./clientError";
 export { ClientHttpError } from "./http/httpError";
+export { handleErrorResponse } from "./http/handleErrorResponse";
 export {
   DEFAULT_TYPE,
   PROBLEM_DETAILS_MIME,
@@ -34,3 +35,63 @@ export {
   type WithErrorResponse,
   hasErrorResponse,
 } from "./http/errorResponse";
+export {
+  BadRequestError,
+  BAD_REQUEST_STATUS,
+  type BadRequestErrorResponse,
+} from "./http/wellKnown/badRequestError";
+export {
+  ConflictError,
+  CONFLICT_STATUS,
+  type ConflictErrorResponse,
+} from "./http/wellKnown/conflictError";
+export {
+  ForbiddenError,
+  FORBIDDEN_STATUS,
+  type ForbiddenErrorResponse,
+} from "./http/wellKnown/forbiddenError";
+export {
+  GoneError,
+  GONE_STATUS,
+  type GoneErrorResponse,
+} from "./http/wellKnown/goneError";
+export {
+  InternalServerError,
+  INTERNAL_SERVER_ERROR_STATUS,
+  type InternalServerErrorResponse,
+} from "./http/wellKnown/internalServerError";
+export {
+  MethodNotAllowedError,
+  METHOD_NOT_ALLOWED_STATUS,
+  type MethodNotAllowedErrorResponse,
+} from "./http/wellKnown/methodNotAllowedError";
+export {
+  NotAcceptableError,
+  NOT_ACCEPTABLE_STATUS,
+  type NotAcceptableErrorResponse,
+} from "./http/wellKnown/notAcceptableError";
+export {
+  NotFoundError,
+  NOT_FOUND_STATUS,
+  type NotFoundErrorResponse,
+} from "./http/wellKnown/notFoundError";
+export {
+  PreconditionFailedError,
+  PRECONDITION_FAILED_STATUS,
+  type PreconditionFailedErrorResponse,
+} from "./http/wellKnown/preconditionFailedError";
+export {
+  TooManyRequestsError,
+  TOO_MANY_REQUESTS_STATUS,
+  type TooManyRequestsErrorResponse,
+} from "./http/wellKnown/tooManyRequestsError";
+export {
+  UnauthorizedError,
+  UNAUTHORIZED_STATUS,
+  type UnauthorizedErrorResponse,
+} from "./http/wellKnown/unauthorizedError";
+export {
+  UnsupportedMediaTypeError,
+  UNSUPPORTED_MEDIA_TYPE_STATUS,
+  type UnsupportedMediaTypeErrorResponse,
+} from "./http/wellKnown/unsupportedMediaTypeError";